+++ app/src/main/java/com/example/friendlygram/MainActivity.kt
@@ -2,23 +2,10 @@ package com.example.friendlygram
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.os.CancellationSignal
-import android.view.View
-import android.widget.Toast
 import androidx.appcompat.widget.Toolbar
 import com.example.friendlygram.databinding.ActivityMainBinding
-import com.example.friendlygram.ui.ChatFragment
-import com.example.friendlygram.ui.SettingsFragment
+import com.example.friendlygram.ui.fragments.ChatFragment
 import com.example.friendlygram.ui.objects.AppDrawer
-import com.mikepenz.materialdrawer.AccountHeader
-import com.mikepenz.materialdrawer.AccountHeaderBuilder
-import com.mikepenz.materialdrawer.Drawer
-import com.mikepenz.materialdrawer.DrawerBuilder
-import com.mikepenz.materialdrawer.model.DividerDrawerItem
-import com.mikepenz.materialdrawer.model.PrimaryDrawerItem
-import com.mikepenz.materialdrawer.model.ProfileDrawerItem
-import com.mikepenz.materialdrawer.model.interfaces.IDrawerItem
-import java.util.function.Consumer
 
 class MainActivity : AppCompatActivity() {
