--- app/src/main/java/com/example/friendlygram/ui/ChatFragment.kt
@@ -1,62 +1,0 @@
-package com.example.friendlygram.ui
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import com.example.friendlygram.R
-import com.example.friendlygram.databinding.FragmentChatBinding
-
-// TODO: Rename parameter arguments, choose names that match
-// the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-//private const val ARG_PARAM1 = "param1"
-//private const val ARG_PARAM2 = "param2"
-//
-
-class ChatFragment : Fragment() {
-
-//    // TODO: Rename and change types of parameters
-//    private var param1: String? = null
-//    private var param2: String? = null
-
-//    override fun onCreate(savedInstanceState: Bundle?) {
-//        super.onCreate(savedInstanceState)
-//        arguments?.let {
-//            param1 = it.getString(ARG_PARAM1)
-//            param2 = it.getString(ARG_PARAM2)
-//        }
-//    }
-
-    private lateinit var mBinding: FragmentChatBinding
-
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
-                              savedInstanceState: Bundle?): View? {
-        mBinding = FragmentChatBinding.inflate(layoutInflater)
-
-        return mBinding.root
-    }
-
-    override fun onResume() {
-        super.onResume()
-    }
-
-    //    companion object {
-//        /**
-//         * Use this factory method to create a new instance of
-//         * this fragment using the provided parameters.
-//         *
-//         * @param param1 Parameter 1.
-//         * @param param2 Parameter 2.
-//         * @return A new instance of fragment ChatFragment.
-//         */
-//        // TODO: Rename and change types and number of parameters
-//        @JvmStatic fun newInstance(param1: String, param2: String) =
-//                ChatFragment().apply {
-//                    arguments = Bundle().apply {
-//                        putString(ARG_PARAM1, param1)
-//                        putString(ARG_PARAM2, param2)
-//                    }
-//                }
-//    }
-}
