+++ app/src/main/java/com/example/friendlygram/ui/fragments/ChatFragment.kt
@@ -1,0 +1,12 @@
+package com.example.friendlygram.ui.fragments
+
+import com.example.friendlygram.R
+
+
+class ChatFragment : BaseFragment(R.layout.fragment_chat) {
+
+    override fun onResume() {
+        super.onResume()
+    }
+
+}
