+++ app/src/main/java/com/example/friendlygram/ui/fragments/SettingsFragment.kt
@@ -1,0 +1,13 @@
+package com.example.friendlygram.ui.fragments
+
+import com.example.friendlygram.R
+
+
+class SettingsFragment : BaseFragment(R.layout.fragment_settings) {
+
+
+    override fun onResume() {
+        super.onResume()
+    }
+
+    }
