+++ app/src/main/java/com/example/friendlygram/ui/objects/AppDrawer.kt
@@ -3,9 +3,8 @@ package com.example.friendlygram.ui.objects
 import android.view.View
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.Toolbar
-import com.example.friendlygram.MainActivity
 import com.example.friendlygram.R
-import com.example.friendlygram.ui.SettingsFragment
+import com.example.friendlygram.ui.fragments.SettingsFragment
 import com.mikepenz.materialdrawer.AccountHeader
 import com.mikepenz.materialdrawer.AccountHeaderBuilder
 import com.mikepenz.materialdrawer.Drawer
