+++ app/src/main/res/layout/fragment_chat.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.ChatFragment">
+    tools:context=".ui.fragments.ChatFragment">
 
     <!-- TODO: Update blank fragment layout -->
